requirements.txt
alembic==1.14.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.9.0
dacite==1.8.1
dask==2023.6.1
distributed==2023.6.1
fsspec==2023.6.0
greenlet==3.1.1
ImageHash==4.3.1
joblib==1.4.2
lightgbm==4.5.0
llvmlite==0.40.1
locket==1.0.0
Mako==1.3.9
MarkupSafe==3.0.2
multimethod==1.9.1
networkx==3.1
numba==0.57.1
numpy==1.24.4
optuna==4.2.1
packaging==24.2
pandas==2.2.3
partd==1.4.0
patsy==0.5.3
phik==0.12.3
psutil==7.0.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
SQLAlchemy==2.0.38
statsmodels==0.14.0
stumpy==1.11.1
threadpoolctl==3.5.0
tqdm==4.67.1
tsfresh==0.20.1
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
visions==0.7.5
wordcloud==1.9.2
ydata-profiling==4.5.1
